# Conflicts:
#	GitTfs.VsCommon/TfsHelper.Common.cs
#	GitTfs/Commands/Fetch.cs
#	GitTfs/Commands/InitBranch.cs
#	GitTfs/ConfigProperties.cs
#	GitTfs/Core/DerivedGitTfsRemote.cs
#	GitTfs/Core/GitTfsRemote.cs
#	GitTfs/Core/IGitTfsRemote.cs
#	GitTfs/Core/TfsInterop/ITfsHelper.cs
#	GitTfs/GitTfs.csproj
#	GitTfs/GitTfsConstants.cs

+++ GitTfs/Commands/CheckinBase.cs
@@ -59,6 +59,6 @@ namespace Sep.Git.Tfs.Commands
             return GitTfsExitCodes.OK;
         }
 
-        protected abstract long DoCheckin(TfsChangesetInfo changeset, string refToCheckin);
+        protected abstract int DoCheckin(TfsChangesetInfo changeset, string refToCheckin);
     }
 }
